render client component for not found
